+++ packages/shadcn/test/CheckboxesWidget.test.tsx
@@ -1,35 +1,32 @@
+import { render } from '@testing-library/react';
+
 import CheckboxesWidget from '../src/CheckboxesWidget';
-import renderer from 'react-test-renderer';
 import { makeWidgetMockProps } from './helpers/createMocks';
 
 describe('CheckboxesWidget', () => {
   test('simple', () => {
-    const tree = renderer
-      .create(
-        <CheckboxesWidget
-          {...makeWidgetMockProps({
-            options: {
-              enumOptions: [{ label: 'A', value: 'a' }],
-            },
-          })}
-        />,
-      )
-      .toJSON();
-    expect(tree).toMatchSnapshot();
+    const { asFragment } = render(
+      <CheckboxesWidget
+        {...makeWidgetMockProps({
+          options: {
+            enumOptions: [{ label: 'A', value: 'a' }],
+          },
+        })}
+      />,
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
   test('inline', () => {
-    const tree = renderer
-      .create(
-        <CheckboxesWidget
-          {...makeWidgetMockProps({
-            options: {
-              enumOptions: [{ label: 'A', value: 'a' }],
-              inline: true,
-            },
-          })}
-        />,
-      )
-      .toJSON();
-    expect(tree).toMatchSnapshot();
+    const { asFragment } = render(
+      <CheckboxesWidget
+        {...makeWidgetMockProps({
+          options: {
+            enumOptions: [{ label: 'A', value: 'a' }],
+            inline: true,
+          },
+        })}
+      />,
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 });